+++ MK4duo/src/lcd/ultralcd.h
@@ -46,6 +46,12 @@
   void lcd_kill_screen();
   void kill_screen(const char* lcd_msg);
 
+  #if (ENABLED(LCD_I2C_TYPE_MCP23017) || ENABLED(LCD_I2C_TYPE_MCP23008)) && ENABLED(DETECT_DEVICE)
+    bool lcd_detected();
+  #else
+    inline bool lcd_detected() { return true; }
+  #endif
+
   #if HAS(BUZZER)
     void lcd_buzz(long duration, uint16_t freq);
   #endif